Support qualification/b-final in rounds

+++ WcaOnRails/app/models/competition_event.rb
@@ -43,6 +43,11 @@ class CompetitionEvent < ApplicationRecord
   end
 
   def load_wcif!(wcif)
+    if self.rounds.pluck(:old_type).compact.any?
+      raise WcaExceptions::BadApiParameter.new(
+        "Cannot edit rounds for a competition which has qualification rounds or b-finals. Please contact WRT or WST if you need to make change to this competition.",
+      )
+    end
     self.rounds.destroy_all!
     total_rounds = wcif["rounds"].size
     wcif["rounds"].each_with_index do |wcif_round, index|
